+++ bcolz/__init__.py
@@ -72,7 +72,7 @@ except ImportError:
               "If on Python2.6 please install unittest2")
 
 
-def get_git_descrtiption(path_):
+def _get_git_descrtiption(path_):
     """ Get the output of git-describe when executed in a given path. """
 
     # imports in function because:
@@ -99,7 +99,7 @@ def get_git_descrtiption(path_):
     except subprocess.CalledProcessError:  # not in git repo
         pass
 
-git_description = get_git_descrtiption(__path__[0])
+git_description = _get_git_descrtiption(__path__[0])
 
 # Initialization code for the Blosc and numexpr libraries
 _blosc_init()