+++ RELEASE_NOTES.rst
@@ -5,7 +5,7 @@ Release notes for bcolz
 Changes from 0.7.3 to 0.7.4
 ===========================
 
-- pass
+- Other miscellaneous fixes and improvements
 
 Changes from 0.7.2 to 0.7.3
 ===========================